pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-society"
version = "0.1.0"
description = "2,500-Agent LLM-driven Society Simulation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "LLM Society Team"}
]
keywords = ["llm", "agent", "simulation", "society", "mesa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mesa>=1.0",
    "numpy>=1.21",
    "scipy>=1.7",
    "networkx>=2.6",
    "pandas>=1.3",
    "sqlalchemy>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
gpu = [
    "pyflamegpu",
    "pycuda",
    "cupy",
]
llm = [
    "google-generativeai",
    "openai",
    "anthropic",
]
visualization = [
    "matplotlib>=3.5",
    "plotly>=5.0",
    "streamlit>=1.20",
]
all = [
    "llm-society[dev,llm,visualization]",
]

[project.scripts]
llm-society = "llm_society.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.setuptools.packages.find]
where = ["."]
include = ["llm_society*"]
